refactor: JpaPlaceOneLineReviewStorageGateway 삭제 및 관련 로직 JpaPlaceStorageGateway로 통합 (#70)
