+++ src/main/kotlin/kr/wooco/woocobe/place/domain/gateway/PlaceStorageGateway.kt
@@ -1,6 +1,7 @@
 package kr.wooco.woocobe.place.domain.gateway
 
 import kr.wooco.woocobe.place.domain.model.Place
+import kr.wooco.woocobe.place.domain.model.PlaceOneLineReviewStat
 
 interface PlaceStorageGateway {
     fun save(place: Place): Place
@@ -10,4 +11,6 @@ interface PlaceStorageGateway {
     fun getByKakaoMapPlaceId(kakaoMapPlaceId: String): Place?
 
     fun getAllByPlaceIds(placeIds: List<Long>): List<Place>
+
+    fun getOneLineReviewStats(placeId: Long): List<PlaceOneLineReviewStat>
 }
